requirements_utf8.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.1
Django==5.2.4
django-allauth==65.11.2
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
django-debug-toolbar==4.4.6
black==24.10.0
flake8==7.1.1
pre-commit==4.0.1
whitenoise==6.6.0
django-redis==5.4.0
